requirements.txt
flask==3.1.0
flask-socketio==5.5.1
feedparser>=6.0.11
apscheduler>=3.10.4
openai>=1.0.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
smartapi-python>=1.4.0
pyotp>=2.9.0
requests>=2.31.0
logzero>=1.7.0
websocket-client>=1.6.0
